requirements.txt
pyautogui
numpy
Pillow
yeelight
pywin32